se2021-23t1-einvoicing-api-h13a-sox-sending-api/requirements.txt
blinker==1.5
cffi==1.15.1
click==8.1.3
cryptography==40.0.1
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Mail==0.9.1
flask-swagger-ui==4.11.1
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycparser==2.21
PyJWT==2.6.0
six==1.16.0
Werkzeug==2.2.3
zipp==3.15.0
